Solar_Forecast_24H/requirements.txt
openmeteo_requests
requests-cache
pandas
numpy
retry-requests
pytz
python-dotenv
pysolar
openpyxl
joblib
scikit-learn
xgboost
requests